\name{init}
\alias{init}
\title{
Launch ShinyApp for Illumina HumanMethylationEPIC manifest
}
\description{
Initialises the launch of a manifest of the EPIC array for convenient browsing of manifest.
}
\usage{
init()
}

\author{
Tyler Gorrie-Stone
Who to contact: <tgorri@essex.ac.uk>
}

\examples{
# init()
}
